backend/app/requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aerich==0.4.3             # via -r requirements.in
aiosqlite==0.16.0         # via tortoise-orm
asyncio-redis==0.16.0     # via broadcaster
asyncpg==0.21.0           # via tortoise-orm
broadcaster[redis]==0.2.0  # via -r requirements.in
click==7.1.2              # via aerich, pip-tools, uvicorn
ddlparse==1.9.0           # via aerich
fastapi==0.63.0           # via -r requirements.in
h11==0.12.0               # via uvicorn
iso8601==0.1.13           # via tortoise-orm
jinja2==2.11.2            # via -r requirements.in
markupsafe==1.1.1         # via jinja2
pip-tools==5.5.0          # via -r requirements.in
pydantic==1.7.3           # via aerich, fastapi
pyparsing==2.4.7          # via ddlparse
pypika==0.44.1            # via tortoise-orm
pytz==2020.5              # via tortoise-orm
starlette==0.13.6         # via fastapi
tortoise-orm[asyncpg]==0.16.19  # via -r requirements.in, aerich
typing-extensions==3.7.4.3  # via aiosqlite
uvicorn==0.13.3           # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
